BEL-899 Move DNS from rails to container component
